Extract alignment and data_parsing modules from the monolith (split PR 2/N)
Independent extraction — does NOT depend on refactor/extract-io-pwm; the
two branches can merge in any order.

- tfbs_footprinter3/alignment.py (new):
    fasta_writer, remove_non_ACGT, remove_gap_only,
    remove_duplicate_species, load_genome_aligned
  (retrieve_genome_aligned and alignment_tools remain in the monolith
  for now because they call ensemblrest; they'll move together with
  the ensembl.py extraction.)

- tfbs_footprinter3/data_parsing.py (new):
    parse_transcript_ids, parse_tf_ids, file_to_datalist,
    compare_tfs_list_jaspar, clean_jaspar_names

- tfbs_footprinter3/tfbs_footprinter3.py: deletes those function defs;
  adds re-export imports so existing callers (tests, hpc/, downstream)
  continue to work via 'from tfbs_footprinter3.tfbs_footprinter3 import X'.

Behavior-neutral: 21 tests green, lint clean, smoke verifies each
re-exported name resolves to a callable.
